pyproject.toml
[tool.poetry]
name = "djangonotesapp"
version = "0.1.0"
description = ""
authors = ["rahulsankrit07@gmail.com <rahulsankrit07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
pytest-django = "^4.5.2"
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

